Change of policy for creation of the libscran meta-library.
We now consider aggregate_across_cells to be a feeder into the
meta-library. Users are not expected to use aggregate_across_cells
directly, either in terms of CMake or the C++ namespace.

As part of this change, we put everything into a namespace
corresponding to the file name. This guarantees that there will not be
any conflicts between files when the meta-library is created.