More comprehensive tests against any assumption of zeroed arrays.
We use a test-only macro to insert a non-zero initial value that changes
with each invocation. So, any reliance on a zeroed input array would
manifest as different results across repeated calls in the test suite.
In effect, all tests are now checking against "dirty" input buffers,
eliminating the need for dedicated tests with less coverage.